--- http4k-security/oauth/src/main/kotlin/org/http4k/security/oauth/core/ClientId.kt
@@ -1,3 +1,0 @@
-package org.http4k.security.oauth.core
-
-data class ClientId(val value: String)
