+++ http4k-security/oauth/src/main/kotlin/org/http4k/security/oauth/server/AccessTokens.kt
@@ -2,7 +2,6 @@ package org.http4k.security.oauth.server
 
 import dev.forkhandles.result4k.Result
 import org.http4k.security.AccessToken
-import org.http4k.security.oauth.core.ClientId
 import org.http4k.security.oauth.server.accesstoken.AuthorizationCodeAccessTokenRequest
 
 /**
