+++ http4k-security/oauth/src/main/kotlin/org/http4k/security/oauth/server/AuthRequest.kt
@@ -6,7 +6,6 @@ import org.http4k.security.ResponseMode
 import org.http4k.security.ResponseType
 import org.http4k.security.ResponseType.Code
 import org.http4k.security.State
-import org.http4k.security.oauth.core.ClientId
 import org.http4k.security.oauth.server.request.RequestObject
 import org.http4k.security.openid.RequestJwtContainer
