+++ http4k-security/oauth/src/main/kotlin/org/http4k/security/oauth/server/OAuthServerMoshi.kt
@@ -16,7 +16,6 @@ import org.http4k.security.Nonce
 import org.http4k.security.ResponseMode
 import org.http4k.security.ResponseType
 import org.http4k.security.State
-import org.http4k.security.oauth.core.ClientId
 
 internal object OAuthServerMoshi : ConfigurableMoshi(
     Moshi.Builder()
