+++ http4k-security/oauth/src/main/kotlin/org/http4k/security/oauth/server/TokenRequest.kt
@@ -1,7 +1,6 @@
 package org.http4k.security.oauth.server
 
 import org.http4k.core.Uri
-import org.http4k.security.oauth.core.ClientId
 import org.http4k.security.oauth.server.accesstoken.GrantType
 import org.http4k.security.oauth.core.RefreshToken
