+++ http4k-security/oauth/src/main/kotlin/org/http4k/security/oauth/server/accesstoken/ClientCredentialsAccessTokenGenerator.kt
@@ -4,7 +4,7 @@ import dev.forkhandles.result4k.map
 import org.http4k.core.Request
 import org.http4k.security.AccessTokenDetails
 import org.http4k.security.oauth.server.AccessTokens
-import org.http4k.security.oauth.core.ClientId
+import org.http4k.security.oauth.server.ClientId
 import org.http4k.security.oauth.server.TokenRequest
 
 class ClientCredentialsAccessTokenGenerator(private val accessTokens: AccessTokens) : AccessTokenGenerator {
