+++ http4k-security/oauth/src/main/kotlin/org/http4k/security/oauth/server/refreshtoken/RefreshTokenAccessTokenGenerator.kt
@@ -6,7 +6,7 @@ import dev.forkhandles.result4k.map
 import org.http4k.core.Request
 import org.http4k.security.AccessTokenDetails
 import org.http4k.security.oauth.server.AccessTokenError
-import org.http4k.security.oauth.core.ClientId
+import org.http4k.security.oauth.server.ClientId
 import org.http4k.security.oauth.server.InvalidRequest
 import org.http4k.security.oauth.server.TokenRequest
 import org.http4k.security.oauth.server.accesstoken.AccessTokenGenerator
