+++ http4k-security/oauth/src/main/kotlin/org/http4k/security/oauth/server/request/RequestObject.kt
@@ -5,7 +5,7 @@ import org.http4k.security.Nonce
 import org.http4k.security.ResponseMode
 import org.http4k.security.ResponseType
 import org.http4k.security.State
-import org.http4k.security.oauth.core.ClientId
+import org.http4k.security.oauth.server.ClientId
 
 data class RequestObject(val client: ClientId? = null,
                          val redirectUri: Uri? = null,
