+++ http4k-security/oauth/src/main/kotlin/org/http4k/security/oauth/server/request/RequestObjectExtractor.kt
@@ -8,7 +8,7 @@ import org.http4k.core.Uri
 import org.http4k.security.ResponseMode
 import org.http4k.security.ResponseType
 import org.http4k.security.State
-import org.http4k.security.oauth.core.ClientId
+import org.http4k.security.oauth.server.ClientId
 import org.http4k.security.oauth.server.InvalidRequestObject
 import org.http4k.security.oauth.server.OAuthServerMoshi
 import org.http4k.security.oauth.server.boolean
