+++ http4k-security/oauth/src/test/kotlin/org/http4k/security/oauth/server/SimpleAuthoriseRequestValidatorTest.kt
@@ -10,7 +10,6 @@ import org.http4k.core.Request
 import org.http4k.core.Uri
 import org.http4k.security.ResponseType
 import org.http4k.security.State
-import org.http4k.security.oauth.core.ClientId
 import org.junit.jupiter.api.Test
 
 class SimpleAuthoriseRequestValidatorTest {
