+++ http4k-security/oauth/src/test/kotlin/org/http4k/security/oauth/server/dummy.kt
@@ -8,7 +8,6 @@ import org.http4k.core.Response
 import org.http4k.core.Uri
 import org.http4k.security.AccessToken
 import org.http4k.security.Nonce
-import org.http4k.security.oauth.core.ClientId
 import org.http4k.security.oauth.server.accesstoken.AuthorizationCodeAccessTokenRequest
 import org.http4k.security.oauth.core.RefreshToken
 import org.http4k.security.oauth.server.refreshtoken.RefreshTokens
