+++ http4k-security/oauth/src/test/kotlin/org/http4k/security/oauth/server/request/RequestObjectExtractorTest.kt
@@ -11,7 +11,7 @@ import org.http4k.format.Jackson
 import org.http4k.security.ResponseMode.Query
 import org.http4k.security.ResponseType.CodeIdToken
 import org.http4k.security.State
-import org.http4k.security.oauth.core.ClientId
+import org.http4k.security.oauth.server.ClientId
 import org.http4k.security.oauth.server.InvalidRequestObject
 import org.http4k.security.oauth.server.request.RequestObjectExtractor.extractRequestJwtClaimsAsMap
 import org.junit.jupiter.api.Test
